+++ lydian/util.py
@@ -9,6 +9,7 @@ import traceback
 from collections.abc import Callable, Generator, Iterable, Mapping, Sequence
 from dataclasses import Field, fields, is_dataclass
 from datetime import UTC, datetime, timedelta, tzinfo
+from inspect import getmembers
 from itertools import zip_longest
 from math import ceil, floor
 from pathlib import Path
@@ -285,6 +286,10 @@ def assure(condition: bool, exc_args: str = '') -> None:  # noqa: FBT001
     if not condition:
         raise AssuranceError(exc_args)
 
+def cog_commands(cog: type[commands.Cog]) -> dict[str, commands.Command]:
+    """Returns a dictionary of a cog's command methods."""
+    return {name:attr for name, attr in getmembers(cog) if isinstance(attr, commands.Command)}
+
 def compose(funcs: Iterable[Callable]) -> Callable[[object], object]:
     """Composes ``funcs`` into one function which takes a single argument.
 
@@ -353,6 +358,10 @@ def get_background_tasks(bot: commands.Bot) -> dict[str, dict[str, tasks.Loop]]:
         cog_name:{name:attr for name, attr in cog.__dict__.items() if isinstance(attr, tasks.Loop)}
         for cog_name, cog in bot.cogs.items()
     }
+
+def getclass[T](obj: type[T] | T) -> type[T]:
+    """Returns ``obj.__class__`` if ``obj`` is not a type, otherwise returns ``obj``."""
+    return obj if isinstance(obj, type) else obj.__class__
 
 def get_dataclass_fields(dc: object, parents: list[str] | None = None) -> dict[str, Field]:
     """Returns a dictionary of field names (dotted if the field is a dataclass) to field objects for a dataclass.